bill_fraud_system/requirements.txt
torch
torchvision
scikit-learn
pillow
numpy
pandas
joblib
